backend/requirements.txt
fastapi
uvicorn
authlib
httpx
python-dotenv
sqlalchemy
pymysql
cryptography
pydantic
itsdangerous
starlette
requests
google-api-python-client
google-auth-oauthlib
python-jose[cryptography]
